Sample Record of detected objects from pi
During this sample run, the pi camera was please in two different states. Lights off and lights on in a room. The model was able to detect certain items in a room with very low lighting. When there was light the model was able to detect objects clearly and more accurately.